+++ src/ShippingFile/Controller.php
@@ -7,9 +7,10 @@
 
 namespace aryelgois\BankInterchange\ShippingFile;
 
+use aryelgois\Utils\Utils;
 use aryelgois\Medools;
+use aryelgois\BankInterchange;
 use aryelgois\BankInterchange\Models;
-use aryelgois\BankInterchange\Utils;
 
 /**
  * Controller class for shipping files
@@ -46,14 +47,16 @@ class Controller
     {
         $shipping_file = Models\ShippingFile::getInstance($where);
 
+        $assignment = $shipping_file->assignment;
+
         $view_class = __NAMESPACE__ . "\\Views\\Cnab$shipping_file->cnab\\"
-            . Utils::toPascalCase($shipping_file->assignment->bank->name);
+            . BankInterchange\Utils::toPascalCase($assignment->bank->name);
 
         $view = new $view_class($shipping_file);
 
         $this->views[] = [
             'file' => $view,
-            'name' => Utils::addExtension(
+            'name' => BankInterchange\Utils::addExtension(
                 $name ?? $view->filename(),
                 static::EXTENSION
             ),
@@ -104,7 +107,10 @@ class Controller
         }
         $zip->close();
 
-        $name = Utils::addExtension($name ?? 'download', '.zip');
+        $name = BankInterchange\Utils::addExtension(
+            $name ?? 'download',
+            '.zip'
+        );
 
         header('Content-Type: application/zip');
         header('Content-Length: ' . filesize($filepath));